fixed additional CRLFs for Source/dolphin-emu.sln
This is a total mess: combination of DOS CRLF and .gitattributes induces inconsistencies depending on how a file
is examined, causing a 100% difference of the file dolphin-emu.sln between master and VR-Hydra branches. This
is despise the fact that the checkout on both branches show DOS format each time. The fix required to convert
the file to Unix so it can be recorded in git history as Unix, then restored as DOS. Even though the file is
checkout as DOS for both branches (as before), they are now both in Unix format in the git history, and the
difference doesn't cover the whole file, allowing a viable merge.